requirements.txt
selenium>=4.0.0
pytest>=7.0.0
allure-pytest>=2.13.5
Faker>=13.0.0 
allure-python-commons~=2.14.3